Move lib into src
Due to typescript structure requirements I have moved the lib folder
into the src foler. I will have to see in the future how a "package" can
be made for just the library. We do not want to export all the server
side code to the client as this is excess size.
