+++ src/lib/shopping.ts
