+++ tsconfig.json
@@ -17,11 +17,7 @@
     // "sourceMap": true,                     /* Generates corresponding '.map' file. */
     // "outFile": "./",                       /* Concatenate and emit output to single file. */
     "outDir": "build" /* Redirect output structure to the directory. */,
-    "rootDir": [
-      "src",
-      "lib",
-      "mocks"
-    ] /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */,
+    "rootDir": "src" /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */,
     // "composite": true,                     /* Enable project compilation */
     // "tsBuildInfoFile": "./",               /* Specify file to store incremental compilation information */
     // "removeComments": true,                /* Do not emit comments to output. */
